拆分OSSRequest
