+++ AliyunOSSSDK/OSSConstants.m
@@ -1,0 +1,15 @@
+//
+//  OSSConstants.m
+//  AliyunOSSSDK
+//
+//  Created by huaixu on 2018/1/22.
+//  Copyright © 2018年 aliyun. All rights reserved.
+//
+
+#import "OSSConstants.h"
+
+NSString * const OSSXMLDictionaryAttributesKey = @"__attributes";
+NSString * const OSSXMLDictionaryCommentsKey = @"__comments";
+NSString * const OSSXMLDictionaryTextKey = @"__text";
+NSString * const OSSXMLDictionaryNodeNameKey = @"__name";
+NSString * const OSSXMLDictionaryAttributePrefix = @"_";
