+++ AliyunOSSSDK/OSSService.h
@@ -11,7 +11,15 @@
 #define OSS_IOS_SDK_VERSION OSSSDKVersion
 
 #import "OSSDefine.h"
+#import "OSSConstants.h"
+
 #import "OSSNetworking.h"
+#import "OSSNetworkingRequestDelegate.h"
+#import "OSSAllRequestNeededMessage.h"
+#import "OSSURLRequestRetryHandler.h"
+#import "OSSHttpResponseParser.h"
+#import "OSSRequest.h"
+
 #import "OSSClient.h"
 #import "OSSModel.h"
 #import "OSSUtil.h"
